+++ apps/nextjs/src/components/post.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 
 import type { RouterOutputs } from "~/utils/api";
 import { api } from "~/utils/api";
-import UpdatePost from "../pages/updatepost";
+import UpdatePost from "./updatepost";
 
 type GetDiscussionOutput =
   RouterOutputs["discussion"]["getDiscussions"][number];
